Refactors Chrono data structure
It is not finished though. I checked the UI code
and refactored a bit only to realize it needs to
be rewritten, for sure.

Right now the data management between the
widgets does not only fail in every possible case,
but it is really really weird.

Needas a rework in the future when the refactor
is near its completion.