CaseInformation: Removes current process info
This is definitely not needed. Having the QProcess
have a parent that does not destroy while we need
the proces available just makes it not go out of scope
when needed in button callbacks and so on.

This way we prevent saving QProcess information
in disk with the rest of the Case information, thus
reducing save file size and problems at loading it
from disk.